+++ entrypoint.sh
@@ -19,6 +19,11 @@ if ! [[ "$CLEAN_PERIOD" =~ ^[0-9]+$ ]]; then
   CLEAN_PERIOD=1800
 fi
 
+if ! [[ "$DECAY_TIME" =~ ^[0-9]+$ ]]; then
+  echo "=> DECAY_TIME not defined, use the default value."
+  DECAY_TIME=30
+fi
+
 if [ "${LOOP}" != "false" ]; then
   LOOP=true
 fi
@@ -27,7 +32,7 @@ if [ "${PRETEND}" != "false" ]; then
   PRETEND=true
 fi
 
-if [ "${DEBUG}" == "0" ]; then
+if [ "${DEBUG}" == "false" ]; then
   unset DEBUG
 fi
 
@@ -37,39 +42,17 @@ trap '{ echo "SIGTERM received, exiting."; exit 0; }' SIGTERM
 while [ 1 ]
 do
 
-  # get all terminated containers
-  TERMINATED=$(docker ps -a -q -f status=exited -f status=dead)
-  if [ $DEBUG ]; then echo -e "Terminated container IDs\n$TERMINATED"; fi
-
-  # handle all terminated containers
-  for CONTAINER_ID in $TERMINATED; do
-    CONTAINER_NAME=$(docker inspect --format='{{(index .Name)}}' $CONTAINER_ID | sed -e 's/^\s*\/\(.*\)$/\1/g')
-    LABELS=$(docker inspect --format='{{.Config.Labels}}' $CONTAINER_ID)
-    # check if container belongs to a stack
-    if [[ ${LABELS} == *"io.rancher.stack.name:"* ]]; then
-      if [ $DEBUG ]; then echo "Keeping terminated container $CONTAINER_NAME since it is contained in a stack"; fi
-    else
-      if [ $DEBUG ]; then echo "Removing terminated standalone container $CONTAINER_NAME"; fi
-      [ "${PRETEND}" == "true" ] || docker rm $CONTAINER_NAME
-    fi
-  done
-  unset CONTAINER_ID
+  # prune terminated containers without rancher stack exited longer than DECAY_TIME
+  [ -n "$DEBUG" ] && echo "pruning terminated containers without rancher stack label exited longer than $DECAY_TIME minutes"
+  [ "${PRETEND}" == "true" ] || docker container prune -f --filter "label!=io.rancher.stack.name" --filter "until=${DECAY_TIME}m"
   
-  # remove dangling volumes
-  VOLUMES=$(docker volume ls -qf dangling=true)
-  if [ $DEBUG ]; then 
-    echo "Removing dangling volumes"
-    echo $VOLUMES
-  fi
-  [ "${PRETEND}" == "true" ] || echo $VOLUMES | xargs -r docker volume rm
+  # prune unused volumes
+  [ -n "$DEBUG" ] && echo "pruning dangling volumes"
+  [ "${PRETEND}" == "true" ] || docker volume prune -f
 
   # remove dangling images
-  IMAGES=$(docker images -qf dangling=true)
-  if [ $DEBUG ]; then 
-    echo "Removing dangling images"
-    echo $IMAGES
-  fi
-  [ "${PRETEND}" == "true" ] || echo $IMAGES | xargs -r docker rmi
+  [ -n "$DEBUG" ] && echo "pruning dangling images"
+  [ "${PRETEND}" == "true" ] || docker image prune -f
 
   # Run forever or exit after the first run depending on the value of $LOOP
   [ "${LOOP}" == "true" ] || break