+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* DECAY_TIME specifies time in minutes for wich a container must have been dead before it gets pruned
+* Use docker-ce cli-api (docker container|volume|image prune) with filtering facilities
+
 **1.0** 2017-08-16
 
 * Clean all terminated standalone containers, dangling volumes and dangling images